% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{boxPlot}
\alias{boxPlot}
\title{Box plots for inferred treatment effects of individuals from
\link[teff]{target}}
\usage{
boxPlot(
  x,
  labs = c("Treatment effect", "Outcome", "Treatment", "Not treated", "Treated"),
  lg = NULL
)
}
\arguments{
\item{x}{object of class \code{tarteff}}

\item{labs}{string of characters for the labels of the plot, it refers in order to labels to use for:
Treatment effect, Outcome, Treatment, and levels of the treatment like: Treated and not Treated.}

\item{lg}{position of the legend for treatment levels such as "topright"}
}
\value{
A plot on the current graphics device#'
}
\description{
Box plots for inferred treatment effects of individuals from
\link[teff]{target}
}
\examples{
data(tcell)
homologous<- matrix(c("DDX3Y","DDX3X","KDM5D","KDM5C","PRKY","PRKX","RPS4Y1","RPS4X","TXLNGY", "TXLNG", "USP9Y", "USP9X", "XIST", "XIST", "TSIX", "TSIX"), nrow=2)
pf <- predicteff(tcell, featuresinf=homologous)
res <- target(tcell, pf, effect="positiveandnegative", featuresinf=homologous, nmcov="age", model="log2")
boxplot(res, lg="topright")
}
