% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{data4teff}
\alias{data4teff}
\title{Data with feature (transcriptpmic) and treatment-effect response data
for targeting individuals with a profile associated with high treatment
response, as derived in an independent study with predictteff.}
\format{
list with fields: features and teffdata
}
\usage{
data(data4teff)
}
\description{
Data with feature (transcriptpmic) and treatment-effect response data
for targeting individuals with a profile associated with high treatment
response, as derived in an independent study with predictteff.
}
\keyword{datasets}
