% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feateff.R
\name{feateff}
\alias{feateff}
\title{Extracts treatment, the outcome on which the effect is assessed, and feature data from transcriptomic and methylomic studies}
\usage{
feateff(
  set,
  tname,
  effname,
  reft = NULL,
  refeff = NULL,
  event = NULL,
  covnames = NULL,
  covtype = NULL,
  sva = FALSE,
  betas = TRUE,
  UsegeneSymbol = FALSE
)
}
\arguments{
\item{set}{\code{GenomicRatioSet}, \code{eSet} derived object or
\code{SummarizedExperiment}}

\item{tname}{\code{character} with name of treatment variable}

\item{effname}{\code{character} with name of outcome variable on which the effect is measured (time to event in the case of survival)}

\item{reft}{\code{character} vector that indicates no-treatment and treatment levels (Default: NULL)}

\item{refeff}{\code{character} vector that indicates no-outcome and outcome levels when the outcome on which the effect is measured is categorical (Default: NULL)}

\item{event}{\code{character} with name of event variable in the case of a survival outcome (Default: NULL)}

\item{covnames}{\code{character} vector with names of covariates (Default: NULL)}

\item{covtype}{\code{character} vector with character "n" indicates which covariates are numerical e.g. c("n", "n", "c") (Default: NULL)}

\item{sva}{\code{logical} indicates whether surrogate variable should be added as covariates (Default: NULL)}

\item{betas}{\code{logical} indicates whether beta values be
used if \code{set} is a \code{GenomicRatioSet} (Default: TRUE)}

\item{UsegeneSymbol}{\code{logical} indicates whether genenames should be used as feature names (Default: FALSE)}
}
\description{
Extracts treatment, the outcome on which the effect is assessed, and feature data from transcriptomic and methylomic studies
}
\details{
This function extracts feature and treatment-effects data,
from  \code{eSet} or \code{SummarizedExperiment} objects
for profiling with \link[teff]{profile}. The function includes the option of adding
surrogate variables as additional covariates to the treatment-effect data.
}
\examples{

\donttest{
library(GEOquery)
gsm <- getGEO("GSE17755")
gsm <- gsm[[1]]

#in this example we use sex as the treatment variable do detect groups of
# high sexual dimorphism in arthritis disease.

data4teff <- feateff(gsm, tname="gender:ch1", reft=c("male", "female"),
                     effname="disease:ch1", refeff=c("healthy","arthritis"),
                     covnames="age:ch1", covtype="n",
                     sva=TRUE, UsegeneSymbol=TRUE)

}
}
