% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mean_ci}
\alias{mean_ci}
\title{returns the \code{mean} and the error limits defined by the
\code{confidence interval}.}
\usage{
mean_ci(x, ci = 0.95, error.limit = "both")
}
\arguments{
\item{x}{vector}

\item{ci}{confidence limit}

\item{error.limit}{error}
}
\value{
vector
}
\description{
returns the \code{mean} and the error limits defined by the
\code{confidence interval}.
}
