% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{pasiw12}
\alias{pasiw12}
\title{Response to Brodalumab treatment of psoriasis. The data is used to compare
the prediction of individual treatment response as inferred by predictteff on
psoriasis data. It is available through data(psoriasis)}
\format{
vector
}
\usage{
data(psoriasis)
}
\description{
Response to Brodalumab treatment of psoriasis. The data is used to compare
the prediction of individual treatment response as inferred by predictteff on
psoriasis data. It is available through data(psoriasis)
}
\keyword{datasets}
