% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pteff.R
\name{plot.pteff}
\alias{plot.pteff}
\title{Plots inferred treatment effects of individuals from
\link[teff]{profile}}
\usage{
\method{plot}{pteff}(x, lb = "Associated treatment effect", ...)
}
\arguments{
\item{x}{object of class \code{pteff}}

\item{lb}{label of the y axis for treatment effect.}
}
\value{
A plot on the current graphics device.
}
\description{
Plots inferred treatment effects of individuals from
\link[teff]{profile}
}
\examples{
data(tcell)
homologous<- matrix(c("DDX3Y","DDX3X","KDM5D","KDM5C","PRKY","PRKX","RPS4Y1","RPS4X","TXLNGY", "TXLNG", "USP9Y", "USP9X", "XIST", "XIST", "TSIX", "TSIX"), nrow=2)
pf <- profile(tcell, featuresinf=homologous)
plot(pf)
}
