% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.tarteff}
\alias{plot.tarteff}
\title{Plots inferred treatment effects of individuals from
\link[teff]{target}}
\usage{
\method{plot}{tarteff}(x, ylab = "Effect", lg = NULL, ...)
}
\arguments{
\item{x}{object of class \code{tarteff}}
}
\value{
A plot on the current graphics device.
}
\description{
Plots inferred treatment effects of individuals from
\link[teff]{target}
}
\examples{
data(tcell)
homologous<- matrix(c("DDX3Y","DDX3X","KDM5D","KDM5C","PRKY","PRKX","RPS4Y1","RPS4X","TXLNGY", "TXLNG", "USP9Y", "USP9X", "XIST", "XIST", "TSIX", "TSIX"), nrow=2)
pf <- predicteff(tcell, featuresinf=homologous)
res <- target(tcell, pf, effect="highandlow", featuresinf=homologous, nmcov="age", model="log2")
plot(res, lg="topright")
}
