% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plotPredict}
\alias{plotPredict}
\title{Plots predicted treatment effects of individuals from
\link[teff]{predicteff}}
\usage{
plotPredict(
  x,
  ...,
  rk = NULL,
  lb = "Associated treatment effect",
  xlab = "Subject Ranking",
  ctrl.plot = NULL
)
}
\arguments{
\item{x}{object of class \code{pteff}}

\item{rk}{object of class \code{vector}, if null treatments are
plotted against their ranking, if not then they ara plotted against rk values.}

\item{lb}{label of the y axis for treatment effect.}

\item{xlab}{label of the x axis.}

\item{ctrl.plot}{controls plot legends (when NULL then internally ctrl.plot <- list(lb=c("Not treated", "Treated"),
wht="bottomleft", whs = "topright"))}
}
\value{
A plot on the current graphics device.
data(tcell)
homologous<- matrix(c("DDX3Y","DDX3X","KDM5D","KDM5C","PRKY","PRKX","RPS4Y1","RPS4X","TXLNGY", "TXLNG", "USP9Y", "USP9X", "XIST", "XIST", "TSIX", "TSIX"), nrow=2)
pf <- predicteff(tcell, featuresinf=homologous)
plotPredict(pf)
}
\description{
Plots predicted treatment effects of individuals from
\link[teff]{predicteff}
}
