% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plotTarget}
\alias{plotTarget}
\title{Plots inferred treatment effects of individuals from
\link[teff]{target}}
\usage{
plotTarget(
  x,
  ...,
  labs = c("Group classification", "Effect", "Treatment", "Not treated", "Treated")
)
}
\arguments{
\item{x}{object of class \code{tarteff}}

\item{labs}{string of characters for the labels of the plot, it refers in order to labels to use for:
Group classification, Effect, Treatment, and levels of the treatment like: Treated and not Treated.}
}
\value{
A plot on the current graphics device
data(tcell)
homologous<- matrix(c("DDX3Y","DDX3X","KDM5D","KDM5C","PRKY","PRKX","RPS4Y1","RPS4X","TXLNGY", "TXLNG", "USP9Y", "USP9X", "XIST", "XIST", "TSIX", "TSIX"), nrow=2)
pf <- predicteff(tcell, featuresinf=homologous)
res <- target(tcell, pf, effect="positiveandnegative", featuresinf=homologous, nmcov="age", model="log2")
plotTarget(res)
}
\description{
Plots inferred treatment effects of individuals from
\link[teff]{target}
}
