% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.pteff}
\alias{print.pteff}
\title{Prints pteff object}
\usage{
\method{print}{pteff}(x, ...)
}
\arguments{
\item{x}{object of class \code{pteff}}

\item{...}{parameters for printing}
}
\value{
object of class \code{pteff}
}
\description{
Prints pteff object
}
