% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.tarteff}
\alias{print.tarteff}
\title{Prints tarteff object}
\usage{
\method{print}{tarteff}(x)
}
\arguments{
\item{x}{object of class \code{tarteff}}
}
\value{
object of class \code{tarteff}
}
\description{
Prints tarteff object
}
