% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{Profiles of individuals with significantly high and low treatment effects}
\usage{
profile(
  x,
  featuresinf = NULL,
  cores = 1,
  seed = 1234,
  plot.overlap = FALSE,
  quant = Inf
)
}
\arguments{
\item{x}{a \code{list} with a fields \code{teffdata} and \code{features}.
\code{teffdata} the data for treatment, effect and covariates across
subjects. \code{features} data is the data on which the profiling
is done. Adjusted features for the variables on the \code{teffdata} are
used to fit the forest and extract the profiles of individuals with
significant  treatments.}

\item{featuresinf}{a \code{vector} of characters with the names of the
features to be used; a \code{matrix} of characters whose columns are
names of features whose values will be averaged (Default:NULL).}

\item{cores}{an integer with the number of cores for parallel
computation (Default:1)}

\item{seed}{and integer with the random seed for splitting data into
train (80\%) and test (20\%) sets.}

\item{plot.overlap}{a logical. If \code{TRUE} then plots the overlap of
adjusted feature data across treatments. Parameter available for less
than 20 features (Default: \code{FALSE}).}

\item{quant}{a number from 1 to the number of features that selects the
number of features with top information score from the causal forest.
By default it selects all the features (Default: Inf).}
}
\value{
a \code{list} of class \code{pteff} with fields:
\describe{
\item{profile:}{a \code{list} with fields \code{profhigh} and \code{proflow}
that are matrices with binarized feature data for the individuals with
significantly high and low treatment effects, respectively.}
\item{predictions:}{a \code{vector} with the estimated treatment effect
of the individuals in the test set.}
\item{featurenames:}{a \code{vector} with the names of the features used.}
\item{cl:}{a \code{vector} with the lower limit of the 95\% confidence
intervals for the estimated treatment effect.}
\item{cu:}{a \code{vector} with the upper limit of the 95\% confidence
intervals for the estimated treatment effect.}
\item{subsids:}{a \code{vector} with ids of subjects in the test set.}
\item{treatment:}{a \code{vector} with treatment effect in the test set.}
}
}
\description{
Profiles of individuals with significantly high and low treatment effects
}
\details{
This function sets up feature and treatment-effects data, fits
random causal forest and identify the individuals with significant
treatment effects. Individuals with significant treatment effects
are considered for those whose confidence intervals for the treatment
estimate do not overlap 0.

The function randomly splits the data into test (20\%) and train (80\%)
sets to fit the forest.
}
\examples{
data(tcell)
homologous<- matrix(c("DDX3Y","DDX3X","KDM5D","KDM5C","PRKY","PRKX","RPS4Y1","RPS4X","TXLNGY", "TXLNG", "USP9Y", "USP9X", "XIST", "XIST", "TSIX", "TSIX"), nrow=2)
profile(tcell, featuresinf=homologous)

}
