% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{psoriasis}
\alias{psoriasis}
\title{Data example for predictteff, which is used to infer individual Brodalumab
treatment response.}
\format{
list with fields: features and teffdata
}
\usage{
data(psoriasis)
}
\description{
Data example for predictteff, which is used to infer individual Brodalumab
treatment response.
}
\keyword{datasets}
