% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{targetprofile}
\alias{targetprofile}
\title{Match subjects to profiles}
\usage{
targetprofile(binfeatures, profileRef, match = 0.8)
}
\arguments{
\item{binfeatures}{binarized feature data with respect to the mean value over
the subjects in the feature data to be targeted}

\item{profileRef}{reference profile of positive or negative treatment effects as obtained by
\link[teff]{predicteff}}

\item{match}{\code{numeric} value between 0 and 1 indicating the level of the match for the targeing
across all the binarized features in the profile. A value of 1 means that all the binarized
features must take identical values in the profile (i,e "positive") for an individual to be clsssified
in its subpopulation (of positive treatment effects).}
}
\value{
a logical vector indicating the features that matched the profile
}
\description{
Match subjects to profiles
}
\keyword{internal}
