% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{tcell}
\alias{tcell}
\title{Data example for predictteff, which is used to infer individual sex effect
on tcell abundance in blood.}
\format{
list with fields: features and teffdata
}
\usage{
data(tcell)
}
\description{
Data example for predictteff, which is used to infer individual sex effect
on tcell abundance in blood.
}
\keyword{datasets}
