% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teff.R
\docType{package}
\name{teff}
\alias{teff}
\title{teff (treatment effects classification): Package for classifying individuals into
subpoulations associated with high and low treatment effects using feature data.}
\description{
\code{teff} is a package designed to extract the profiles of subpopulations
with associated high and low treatment effects. With the extracted profiles,
new individuals with feature data can be targeted and classified. If
treatment and effect data is available for these new individuals the package can test
whether the association between the treatment and the effect is indeed different
across subpopulations.
}
