% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Targets individuals into groups of high, low or neutral treatment effects}
\usage{
target(
  x,
  pteffObject,
  effect = "high",
  featuresinf = NULL,
  plot = TRUE,
  model = "gaussian",
  match = 0.8,
  nmcov = NULL,
  cores = 1,
  ...
)
}
\arguments{
\item{x}{a \code{list} with a fields \code{teffdata} and \code{features}.
\code{teffdata} the data for treatment, effect and covariates across
subjects. \code{features} data is the data on which the profiling
is done. Adjusted features for the variables on the \code{teffdata} are
used to fit the forest and extract the profiles of individuals with
significant  treatments.}

\item{pteffObject}{object of class \code{pteff}}

\item{effect}{\code{character} with one of the values: "high", "low", "highandlow", indicating
which subpopulation to be targeted: subpopulation with high treatment effect only, low
treatment effect only or both, respectively (Default: high).}

\item{featuresinf}{a \code{vector} of characters with the names of the
features to be used; a \code{matrix} of characters whose columns are
names of features whose values will be averaged (Default:NULL).}

\item{plot}{\code{logical} indicating a whether a plot of the targeting will be produced}

\item{model}{\code{character} with one of the values corresponding to the family argument of a
general linear model (\link[stats]{glm}), such as: "gaussian", "binomial", etc. It also allows "beta" for beta
regression, "hazard" for proportional hazards model \link[survival]{coxph}, and "log2" for log2
transformation of the effects (Default: "gaussian").}

\item{match}{\code{numeric} value between 0 and 1 indicating the level of the match for the targeing
across all the binarized features in the profile. A value of 1 means that all the binarized
features must take identical values in the profile (i,e "high") for an individual to be clsssified
in its subpopulation (of high treatment effects).}

\item{nmcov}{vector of \code{character} with the names of the covariates to be included in the models.}

\item{cores}{an integer with the number of cores for parallel
computation (Default:1).}

\item{...}{additional parameters for the \code{image} function}
}
\value{
a \code{list} of class \code{tarteff} with fields:
\describe{
\item{classification:}{classification of individuals into subpopulations of
expected high and/or low treatment effects}
\item{summary.model:}{a \code{summary} of the model used to test the interaction between
the profile ("highandlow": high=1, neutral=0, low=-1; "high": high=1, neutral=0,
"low": low=1, neutral=0) with the effect.}
}
}
\description{
Targets individuals into groups of high, low or neutral treatment effects
}
\details{
This function uses feature data of individuals to classify them
into subpopulations associated with high and low treatment effects. The
classification is performed by targeting the feature data (adjusted by
covariates and binarized) to the profiles computed in \link[teff]{profile}.

The function tests whether the classification of the subjects stratifies
the association of the treatment with the effect, testing the
interaction between subpopulation classification and the treatment.

Models on the effects to test the interaction include the general lineal
models, beta regression and proportional hazards models.

The function can be used to target new individuals not used in the profiling
and/or other types of effects expected from the treatment.
}
\examples{
data(tcell)
homologous<- matrix(c("DDX3Y","DDX3X","KDM5D","KDM5C","PRKY","PRKX","RPS4Y1","RPS4X","TXLNGY", "TXLNG", "USP9Y", "USP9X", "XIST", "XIST", "TSIX", "TSIX"), nrow=2)
pf <- predicteff(tcell, featuresinf=homologous, profile=TRUE)
res <- target(tcell, pf, effect="highandlow", featuresinf=homologous, nmcov="age", model="log2")
res

}
